Update README.md
This repository serves as the base for the HUDS Bootcamp for social impact. You can add a deeper description of the project here using the Markdown or LaTex languages. You can format code using the three backticks and close you code in those quotes. Add dependencies for those following you code to be successful in running on their own.